feat: refactor create, update, and delete functionality to functions located in db_functions.php and called in their respective pages.

+++ LearnPHP/MySQL/db_functions.php
@@ -1,4 +1,19 @@
 <?php include "db.php";
+
+  function createUser() {
+    if(isset($_POST['submit'])) {
+      global $connection;
+      $username = $_POST['username'];
+      $password = $_POST['password'];
+      $query = "INSERT INTO users(username,password) ";
+      $query .= "VALUES ('$username', '$password')";
+    
+      $result = mysqli_query($connection, $query);
+    
+      if(!$result) { die('Query FAILED'.mysqli_error()); }
+      else { echo "Record Created!"; }
+    }
+  }
   
   function buildSelectMenu() {
     global $connection;
@@ -15,28 +30,32 @@
   }  
 
   function updateUser() {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if(isset($_POST['submit'])) {
+      global $connection;
+      $username = $_POST['username'];
+      $password = $_POST['password'];
+      $id = $_POST['id'];
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";      $query .= "WHERE id = $id";
+      $query = "UPDATE users SET ";
+      $query .= "username = '$username', ";
+      $query .= "password = '$password' ";      $query .= "WHERE id = $id";
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) { die('Query FAILED '.mysqli_error($connection)); } 
+      $result = mysqli_query($connection, $query);
+      if(!$result) { die('Query FAILED '.mysqli_error($connection)); } 
+    }
   }
 
   function deleteUser() {
-    global $connection;
-    $id = $_POST['id'];
+    if(isset($_POST['submit'])) {
+      global $connection;
+      $id = $_POST['id'];
 
-    $query = "DELETE from users ";
-    $query .= "WHERE id = $id";
+      $query = "DELETE from users ";
+      $query .= "WHERE id = $id";
 
-    $result = mysqli_query($connection, $query);
-    if(!$result) { die('Query FAILED '.mysqli_error($connection)); } 
+      $result = mysqli_query($connection, $query);
+      if(!$result) { die('Query FAILED '.mysqli_error($connection)); } 
+    }  
   }
 
 ?>
